--- Assets/CharakterController.cs
@@ -1,47 +1,0 @@
-﻿using System.Collections;
-using System.Collections.Generic;
-using UnityEngine;
-
-public class CharakterController : MonoBehaviour
-{
-
-    [SerializeField] float speed;                //Floating point variable to store the player's movement speed.
-
-    private Rigidbody2D rb2d;        //Store a reference to the Rigidbody2D component required to use 2D Physics.
-
-    // Use this for initialization
-    void Start()
-    {
-        //Get and store a reference to the Rigidbody2D component so that we can access it.
-        rb2d = GetComponent<Rigidbody2D>();
-    }
-
-    //FixedUpdate is called at a fixed interval and is independent of frame rate. Put physics code here.
-    void FixedUpdate()
-    {
-        faceMouse();
-
-        //Store the current horizontal input in the float moveHorizontal.
-        float moveHorizontal = Input.GetAxis("Horizontal");
-
-        //Store the current vertical input in the float moveVertical.
-        float moveVertical = Input.GetAxis("Vertical");
-
-        //Use the two store floats to create a new Vector2 variable movement.
-        Vector2 movement = new Vector2(moveHorizontal, moveVertical);
-
-        //Call the AddForce function of our Rigidbody2D rb2d supplying movement multiplied by speed to move our player.
-        rb2d.AddForce(movement * speed);
-    }
-
-    void faceMouse()
-    {
-        Vector3 mousepostion = Input.mousePosition;
-        mousepostion = Camera.main.ScreenToWorldPoint(mousepostion);
-
-        Vector2 direction = new Vector2(mousepostion.x - transform.position.x, mousepostion.y - transform.position.y);
-
-        transform.right = direction;
-
-    }
-}
