+++ Assets/PNG.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 1eae61b6015a592448134dde1c247619
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
