+++ Assets/PNG/manRed_silencer.png.meta
