+++ Assets/Scripts/BulletController.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 17b4604463f3ec34691676cefb8d2811
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
